requirements.txt

numpy>=1.21.0
pandas>=1.3.3
matplotlib>=3.4.3
seaborn>=0.11.2
scikit-learn>=1.0.2
tensorflow>=2.9.0
joblib>=1.1.0
pyyaml>=6.0

requests>=2.27.1
python-dotenv>=0.20.0

streamlit>=1.12.0
plotly>=5.6.0
statsmodels>=0.13.2

tqdm>=4.64.0